Improved test structure with shared helpers and functional test separation.
