+++ tests/phpunit/Unit/Prompty/PromptyFlowIntegrationTest.php
@@ -17,34 +17,6 @@ use PHPUnit\Framework\Attributes\Group;
 #[Group('unit')]
 final class PromptyFlowIntegrationTest extends PromptyTestCase {
 
-  /**
-   * Set environment variables for a flow.
-   *
-   * @param array<string, string> $vars
-   *   Key-value pairs to set.
-   * @param string $prefix
-   *   Env prefix.
-   */
-  protected function setEnvVars(array $vars, string $prefix = 'PROMPTY_'): void {
-    foreach ($vars as $key => $value) {
-      putenv($prefix . strtoupper($key) . '=' . $value);
-    }
-  }
-
-  /**
-   * Clear environment variables after a flow.
-   *
-   * @param array<string> $keys
-   *   Keys to clear.
-   * @param string $prefix
-   *   Env prefix.
-   */
-  protected function clearEnvVars(array $keys, string $prefix = 'PROMPTY_'): void {
-    foreach ($keys as $key) {
-      putenv($prefix . strtoupper($key));
-    }
-  }
-
   public function testFlowLinear(): void {
     $this->setEnvVars([
       'name' => 'my-app',
@@ -66,7 +38,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
     $this->assertSame('vue', $result['framework']);
     $this->assertTrue($result['install']);
 
-    $this->clearEnvVars(['name', 'framework', 'install']);
   }
 
   public function testFlowLinearRenderedOutput(): void {
@@ -87,7 +58,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
     $this->assertStringContainsString('+  Framework', $output);
     $this->assertStringContainsString('React', $output);
 
-    $this->clearEnvVars(['name', 'framework']);
   }
 
   public function testFlowNested(): void {
@@ -123,7 +93,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
     $this->assertArrayNotHasKey('lib_format', $result);
     $this->assertTrue($result['api_routes']);
 
-    $this->clearEnvVars(['type', 'app_framework', 'api_routes']);
   }
 
   public function testFlowIntroOutroStrings(): void {
@@ -142,7 +111,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
     $this->assertStringContainsString('#  Welcome', $output);
     $this->assertStringContainsString('#  Done!', $output);
 
-    $this->clearEnvVars(['name']);
   }
 
   public function testFlowIntroOutroCallable(): void {
@@ -172,11 +140,10 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
     $this->assertStringContainsString('Custom intro', $output);
     $this->assertStringContainsString('Custom outro: test', $output);
 
-    $this->clearEnvVars(['name']);
   }
 
   public function testFlowConfig(): void {
-    putenv('MYAPP_NAME=configured');
+    $this->setEnvVars(['name' => 'configured'], 'MYAPP_');
 
     $result = NULL;
     $this->captureOutput(function () use (&$result): void {
@@ -187,8 +154,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
 
     $this->assertNotNull($result);
     $this->assertSame('configured', $result['name']);
-
-    putenv('MYAPP_NAME');
   }
 
   public function testFlowNumbering(): void {
@@ -204,7 +169,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
     $this->assertStringContainsString('(1)', $output);
     $this->assertStringContainsString('(2)', $output);
 
-    $this->clearEnvVars(['name', 'framework']);
   }
 
   public function testFlowNumberingNested(): void {
@@ -224,7 +188,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
     $this->assertStringContainsString('(1)', $output);
     $this->assertStringContainsString('(1.1)', $output);
 
-    $this->clearEnvVars(['type', 'child']);
   }
 
   public function testFlowMultiselectWithEnv(): void {
@@ -242,7 +205,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
     $this->assertNotNull($result);
     $this->assertSame(['ts', 'eslint'], $result['features']);
 
-    $this->clearEnvVars(['features']);
   }
 
   public function testFlowEmptyReturnsEmptyArray(): void {
@@ -265,7 +227,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
 
     $this->assertStringContainsString('Yep', $output);
 
-    $this->clearEnvVars(['install']);
   }
 
   public function testFlowConfigTruthy(): void {
@@ -281,7 +242,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
     $this->assertNotNull($result);
     $this->assertTrue($result['install']);
 
-    $this->clearEnvVars(['install']);
   }
 
   public function testFlowConfigFalsy(): void {
@@ -297,7 +257,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
     $this->assertNotNull($result);
     $this->assertFalse($result['install']);
 
-    $this->clearEnvVars(['install']);
   }
 
   public function testFlowConfigSymbolsAscii(): void {
@@ -311,7 +270,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
 
     $this->assertStringContainsString('*', $output);
 
-    $this->clearEnvVars(['name']);
   }
 
   public function testFlowConfigNoOverrides(): void {
@@ -327,7 +285,6 @@ final class PromptyFlowIntegrationTest extends PromptyTestCase {
     $this->assertNotNull($result);
     $this->assertSame('test', $result['name']);
 
-    $this->clearEnvVars(['name']);
   }
 
   public function testFlowCancelledWithString(): void {
